pyproject.toml
[tool.poetry.scripts]
main = 'src.main:main'

[tool.poetry]
name = "cru-scrape"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
attrs = "23.2.0"
automat = "22.10.0"
beautifulsoup4 = "4.12.3"
bs4 = "0.0.2"
certifi = "2024.7.4"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
constantly = "23.10.4"
cryptography = "43.0.0"
cssselect = "1.2.0"
defusedxml = "0.7.1"
filelock = "3.15.4"
hyperlink = "21.0.0"
idna = "3.7"
incremental = "22.10.0"
itemadapter = "0.9.0"
itemloaders = "1.3.1"
jmespath = "1.0.1"
lxml = "5.2.2"
numpy = "2.0.1"
packaging = "24.1"
pandas = "2.2.2"
parsel = "1.9.1"
protego = "0.3.1"
psycopg2 = "2.9.9"
pyasn1 = "0.6.0"
pyasn1-modules = "0.4.0"
pycparser = "2.22"
pydispatcher = "2.0.7"
pyopenssl = "24.2.1"
python-dateutil = "2.9.0.post0"
pytz = "2024.1"
queuelib = "1.7.0"
requests = "2.32.3"
requests-file = "2.1.0"
scrapy = "2.11.2"
service-identity = "24.1.0"
six = "1.16.0"
soupsieve = "2.5"
tldextract = "5.1.2"
twisted = "24.3.0"
typing-extensions = "4.12.2"
tzdata = "2024.1"
urllib3 = "2.2.2"
w3lib = "2.2.1"
zope-interface = "6.4.post2"
typer = "^0.12.3"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
